refactore property-service

+++ common/common-domain/src/main/java/com/booking/commondomain/dto/property/PageProperties.java
@@ -22,7 +22,7 @@ public class PageProperties {
 
     @NotNull
     @Min(value = 0)
-    private Long totalElements;
+    private int totalElements;
 
     @Valid
     private List<PropertyDto> properties;
